+++ app/models/importers/conversion_employer_plan_year_update.rb
@@ -1,12 +1,5 @@
 module Importers
-  class ConversionEmployerPlanYearUpdate < ConversionEmployerPlanYearCommon
-    validates_length_of :fein, is: 9
-
-    validate :validate_fein
-    validate :validate_new_coverage_policy
-
-    validates_presence_of :plan_selection, :allow_blank => false
-    validates_numericality_of :enrolled_employee_count, :allow_blank => false
+  class ConversionEmployerPlanYearUpdate < ConversionEmployerPlanYear
 
     def calculated_coverage_start
       return @calculated_coverage_start if @calculated_coverage_start
@@ -16,60 +9,6 @@ module Importers
       else
         default_plan_year_start
       end
-    end
-
-    def validate_fein
-      return true if fein.blank?
-      found_employer = find_employer
-      if found_employer.nil?
-        errors.add(:fein, "does not exist")
-      else
-        # if found_employer.plan_years.any? && (found_employer.profile_source == "conversion")
-        #   errors.add(:fein, "employer already has conversion plan years")
-        # end
-      end
-    end
-
-    def validate_new_coverage_policy
-      return true if new_coverage_policy.blank?
-      if new_coverage_policy_value.blank?
-        warnings.add(:new_coverage_policy, "invalid new hire coverage start policy specified (not one of #{HIRE_COVERAGE_POLICIES.keys.join(",")}), defaulting to first of month following date of hire")
-      end
-    end
-
-    def find_employer
-      org = Organization.where(:fein => fein).first
-      return nil unless org
-      org.employer_profile
-    end
-
-    def select_most_common_plan(available_plans, most_expensive_plan)
-        if !most_common_hios_id.blank?
-          mc_hios = most_common_hios_id.strip
-          found_single_plan = available_plans.detect { |pl| (pl.hios_id == mc_hios) || (pl.hios_id == "#{mc_hios}-01") }
-          return found_single_plan if found_single_plan
-          warnings.add(:most_common_hios_id, "hios id #{most_common_hios_id.strip} not found for most common plan, defaulting to most expensive plan")
-        else
-          warnings.add(:most_common_hios_id, "no most common hios id specified, defaulting to most expensive plan")
-        end
-        most_expensive_plan
-    end
-
-    def select_reference_plan(available_plans)
-      plans_by_cost = available_plans.sort_by { |plan| plan.premium_tables.first.cost }
-      most_expensive_plan = plans_by_cost.last
-      if (plan_selection == "single_plan")
-        if !single_plan_hios_id.blank?
-          sp_hios = single_plan_hios_id.strip
-          found_single_plan = available_plans.detect { |pl| (pl.hios_id == sp_hios) || (pl.hios_id == "#{sp_hios}-01") }
-          return found_single_plan if found_single_plan
-          warnings.add(:single_plan_hios_id, "hios id #{single_plan_hios_id.strip} not found for single plan benefit group defaulting to most common plan")
-        else
-          warnings.add(:single_plan_hios_id, "no hios id specified for single plan benefit group, defaulting to most common plan")
-        end
-      end
-
-      select_most_common_plan(available_plans, most_expensive_plan)
     end
 
     def find_and_update_plan_year
@@ -143,23 +82,6 @@ module Importers
     def save 
       return false unless valid?
       find_and_update_plan_year
-    end
-
-    def map_employees_to_benefit_groups(employer, plan_year)
-      bg = plan_year.benefit_groups.first
-      employer.census_employees.non_terminated.each do |ce|
-        ce.add_benefit_group_assignment(bg)
-        ce.save!
-      end
-    end
-
-    def propagate_errors(plan_year)
-      plan_year.errors.each do |attr, err|
-        errors.add("plan_year_" + attr.to_s, err)
-      end
-      plan_year.benefit_groups.first.errors.each do |attr, err|
-        errors.add("plan_year_benefit_group_" + attr.to_s, err)
-      end
     end
   end
 end
